+++ src/core/operators/noblank.ml
@@ -284,7 +284,11 @@ let _ =
           descr = "when detecting a blank.";
           register_deprecated_argument = false;
           arg_t = [];
-          register = (fun ~params:_ s f -> s#on_blank (fun () -> f []));
+          register =
+            (fun ~params:_ s f ->
+              let f, remove = Lang_source.disarmable f in
+              s#on_blank (fun () -> f []);
+              remove);
         };
         {
           name = "on_noise";
@@ -292,7 +296,11 @@ let _ =
           descr = "when noise is detected.";
           register_deprecated_argument = false;
           arg_t = [];
-          register = (fun ~params:_ s f -> s#on_noise (fun () -> f []));
+          register =
+            (fun ~params:_ s f ->
+              let f, remove = Lang_source.disarmable f in
+              s#on_noise (fun () -> f []);
+              remove);
         };
       ]
     (proto frame_t)