+++ src/core/source/start_stop.ml
@@ -130,7 +130,11 @@ let callbacks ~label =
         descr = "when " ^ label ^ " starts";
         register_deprecated_argument = true;
         arg_t = [];
-        register = (fun ~params:_ s f -> s#on_start (fun () -> f []));
+        register =
+          (fun ~params:_ s f ->
+            let f, remove = Lang_source.disarmable f in
+            s#on_start (fun () -> f []);
+            remove);
       };
       {
         name = "on_stop";
@@ -138,7 +142,11 @@ let callbacks ~label =
         descr = "when " ^ label ^ " stops";
         register_deprecated_argument = true;
         arg_t = [];
-        register = (fun ~params:_ s f -> s#on_stop (fun () -> f []));
+        register =
+          (fun ~params:_ s f ->
+            let f, remove = Lang_source.disarmable f in
+            s#on_stop (fun () -> f []);
+            remove);
       };
     ]