+++ src/core/sources/harbor_input.ml
@@ -548,8 +548,10 @@ let callbacks () =
       arg_t = [(false, "", Lang.metadata_t)];
       register =
         (fun ~params:_ s on_connect ->
+          let on_connect, remove = Lang_source.disarmable on_connect in
           let on_connect m = on_connect [("", Lang.metadata_list m)] in
-          s#on_connect on_connect);
+          s#on_connect on_connect;
+          remove);
     };
     {
       name = "on_disconnect";
@@ -557,7 +559,11 @@ let callbacks () =
       descr = "when a source is disconnected.";
       register_deprecated_argument = true;
       arg_t = [];
-      register = (fun ~params:_ s f -> s#on_disconnect (fun () -> f []));
+      register =
+        (fun ~params:_ s f ->
+          let f, remove = Lang_source.disarmable f in
+          s#on_disconnect (fun () -> f []);
+          remove);
     };
   ]